Fixed some actions, and bug with sqlite unsigned int

+++ include/models.hpp
@@ -8,13 +8,13 @@ namespace rs::model {
 
 /* Models for Database */
 struct User : Model {
-    Field<unsigned long,cnstr::Unique> id;
+    Field<int,cnstr::Unique> id;
     Field<std::string, cnstr::Unique, cnstr::Length<1,10>, cnstr::Required> username;
     Field<std::string, cnstr::Required, cnstr::Length<6,255>> password;
     Field<std::string, cnstr::Unique,cnstr::Required, cnstr::NotEmpty, cnstr::Length<2,32>> email;
     Field<std::string, cnstr::Required, cnstr::Length<2,64>> firstname;
     Field<std::string, cnstr::Required, cnstr::Length<2,64>> lastname;
-    Field<unsigned long, cnstr::Required> born;
+    Field<int, cnstr::Required> born;
     Field<std::string> status;
 };
